+++ src/geefetch/data/satellites/__init__.py
@@ -1,6 +1,7 @@
 from .abc import SatelliteABC
 from .dynworld import DynWorld
 from .gedi import GEDIraster, GEDIvector
+from .landsat8 import Landsat8
 from .s1 import S1
 from .s2 import S2
 
@@ -11,4 +12,5 @@ __all__ = [
     "GEDIvector",
     "DynWorld",
     "GEDIraster",
+    "Landsat8",
 ]